+++ Plutarch/List.hs
@@ -57,19 +57,11 @@ module Plutarch.List (
 import Data.Kind (Constraint, Type)
 import GHC.Generics (Generic)
 import Numeric.Natural (Natural)
-import Plutarch.Bool (
-  PBool (PFalse, PTrue),
-  PEq,
-  POrd,
-  pif,
-  (#&&),
-  (#<),
-  (#<=),
-  (#==),
-  (#||),
- )
+import Plutarch.Builtin.Bool (PBool (PFalse, PTrue), pif, (#&&), (#||))
 import Plutarch.Integer (PInteger)
+import Plutarch.Internal.Eq (PEq ((#==)))
 import Plutarch.Internal.Lift (pconstant)
+import Plutarch.Internal.Ord (POrd, (#<), (#<=))
 import Plutarch.Internal.Other (pfix)
 import Plutarch.Internal.PLam (plam)
 import Plutarch.Internal.PlutusType (